4251 removeExtensionForEveryTab
